Fix the connect synchronous function
It was obvious that this was returning true when a connection failed. On initial investigation, this seemed to be because the function was not checking the wlan reason code of a connection complete notification.

On further investigation I realized this reason code was always success, leading me to discover the WLAN_CONNECTION_NOTIFICATION_DATA was not defined correctly. The profile name length was using 32, which is an ssid max length, but not a profile name max length. There was also a missing flag field. After adding these fields, the notification seems to be marshalled correctly now (though I haven't tested the last discovery profile XML field which looks suspicious as well)

After fixing this and checking the reason code in the connectsynchronous function, the function fails correctly now